created billed invoices feature
Billed invoices can be created with

$ rechnung bill-items 2019 10
$ rechnung bill-items 2019 11
$ rechnung bill-items 2019 12

which will create items to be billed for all contracts,
which are active in the given month. Next:

$ rechnung create-billed-invoices 2019.Q4

will create invoices from the billed items. All invoices
will get the suffix 2019.Q4.

Unfortunately this also required a couple of other changes

* contract_data is now passed plain into template. Elements
can now be accessed directly, without "contract."
* invoice_data is now passed plain into template. Elements
can now be accessed directly, withoud "invoice."
* test_fixtures assets folder is symlink to the orig folder,
as we should verify, that the assets we ship work with our tests..